New Installation Scripts

+++ Linux/install_oasys.sh
@@ -1,6 +1,15 @@
 #!/bin/bash
 
-rm -fR $HOME/miniconda3
+type git >/dev/null 2>&1 || sudo ./check_git.sh
+
+read -p "Delete or Rename previous Miniconda3 installation ? [D/r]" -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Rr]$ ]]
+then
+    mv $HOME/miniconda3 $HOME/miniconda3_$(date +'%d-%m-%Y_%H.%M.%S')
+else
+    rm -fR $HOME/miniconda3
+fi
 
 wget https://repo.continuum.io/miniconda/Miniconda3-4.7.12.1-Linux-x86_64.sh
 chmod +x Miniconda3-4.7.12.1-Linux-x86_64.sh
@@ -13,3 +22,5 @@ cd $HOME/miniconda3/bin
 cd -
 
 $HOME/miniconda3/bin/python -m oasys.canvas
+
+sudo ./create_desktop_application.sh
